ground_station/requirements.txt
# AquaBot ground station — Python dependencies
# Install with: pip install -r requirements.txt
#
# Tested on Jetson Orin Nano (JetPack 6.x, Python 3.10+)

flask>=3.0
websockets>=12.0
paho-mqtt>=2.0
pyserial>=3.5
